+++ docs/source/changelog.rst
@@ -4,6 +4,9 @@
 Changelog
 =========
 
+- :release:`3.0.0 <21th April 2023>`
+- :feature:`10` Retrieve package metadata from the JSON API
+
 - :release:`2.0.0 <14th April 2023>`
 - :feature:`4` Use single method for all RSS feeds